Refactoring step 2
