+++ test/util/test_openapi_spec.py
@@ -91,25 +91,3 @@ class TestOpenAPISpecification:
         spec_dict = {"info": {"title": "Test API", "version": "1.0.0"}, "paths": {"/users": {}}}
         with pytest.raises(ValueError):
             OpenAPISpecification(spec_dict)
-
-    #  Should return the raw OpenAPI specification dictionary with resolved references.
-    def test_return_raw_spec_with_resolved_references(self, test_files_path):
-        spec = OpenAPISpecification.from_file(test_files_path / "json" / "complex_types_openapi_service.json")
-        raw_spec = spec.to_dict(resolve_references=True)
-
-        assert "$ref" not in str(raw_spec)
-        assert "#/" not in str(raw_spec)
-
-        # verify that we can serialize the raw spec to a string
-        schema_ser = json.dumps(raw_spec, indent=2)
-
-        # and that the serialized string does not contain any $ref or #/ references
-        assert "$ref" not in schema_ser
-        assert "#/" not in schema_ser
-
-        # and that we can deserialize the serialized string back to a dictionary
-        schema = json.loads(schema_ser)
-        assert "$ref" not in schema
-        assert "#/" not in schema
-
-        assert schema == raw_spec
